// src/entities/Profile/model/types/types.ts
import { Country } from '@/entities/Country';
import { Currency } from '@/entities/Currency';

export interface IProfile {
    id?: string,
    first?: string
    lastname?: string
    age?: number
    currency?: Currency
    country?: Country
    city?: string
    username?: string
    avatar?: string
}